Change to singularis for packages
