+++ extensions/viewevent/src/main/kotlin/smartadapter/viewevent/listener/OnViewEventListener.kt
@@ -1,0 +1,8 @@
+package smartadapter.viewevent.listener
+
+import smartadapter.viewevent.model.ViewEvent
+
+interface OnViewEventListener<T : ViewEvent> {
+
+    var eventListener: (T) -> Unit
+}
