--- extensions/viewevent/src/main/kotlin/smartadapter/viewevent/listeners/OnViewEventListener.kt
@@ -1,8 +1,0 @@
-package smartadapter.viewevent.listeners
-
-import smartadapter.viewevent.models.ViewEvent
-
-interface OnViewEventListener<T : ViewEvent> {
-
-    var eventListener: (T) -> Unit
-}
