+++ extensions/viewevent/src/main/kotlin/smartadapter/viewevent/swipe/BasicSwipeEventBinder.kt
@@ -10,7 +10,7 @@ import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.RecyclerView
 import smartadapter.SmartRecyclerAdapter
 import smartadapter.SmartViewHolderType
-import smartadapter.viewevent.models.ViewEvent
+import smartadapter.viewevent.model.ViewEvent
 import smartadapter.viewholder.OnItemSwipedListener
 import smartadapter.viewholder.SmartViewHolder
 import kotlin.math.abs
