+++ extensions/viewevent/src/main/kotlin/smartadapter/viewevent/viewholder/OnItemMovedEventListener.kt
@@ -1,6 +1,6 @@
 package smartadapter.viewholder
 
-import smartadapter.viewevent.models.ViewEvent
+import smartadapter.viewevent.model.ViewEvent
 
 interface OnItemMovedEventListener {
