+++ extensions/viewevent/src/main/kotlin/smartadapter/viewevent/viewholder/OnItemTouchEventListener.kt
@@ -1,6 +1,6 @@
 package smartadapter.viewholder
 
-import smartadapter.viewevent.models.ViewEvent
+import smartadapter.viewevent.model.ViewEvent
 
 /**
  * Smart adapter item view touch listener.
